moved root container renderer and root container renderer test to the roto the root container folders
