+++ src/electron/views/root-container/root-container-renderer.tsx
